requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.15.2
dnspython==2.7.0
et_xmlfile==2.0.0
gunicorn==21.2.0
MarkupSafe==3.0.2
mongoengine==0.29.1
openpyxl==3.1.5
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
pymongo==3.11.4
pypng==0.20220715.0
python-dotenv==1.1.1
pytz==2025.2
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
whitenoise==6.9.0
